backend/requirements.txt
django
djangorestframework
django-cors-headers
phonenumbers
django-rest-knox